Allow an API user to explicitly target the OBE
This code predates core looking at the feature flag to decide where
to put a dataset when it's not specified, so it assumes that "don't
ask for a backend" implies "put it in the OBE".  That's not true
anymore.  This change keeps the existing semantics unchanged, but
adds a new way to specify the backend (and reimplements the existing
codepath in terms of it).